No data error

+++ scripts/landsat-tile.sh
@@ -98,7 +98,9 @@ for pathrow in "${pathrows[@]}"; do
               "$nbar_angle"
 done
 
-if [[ -f "$nbar_input" ]]; then
+ls "$workingdir"
+stat -c %s "$nbar_input"
+if [[ -f "$nbar_input" ]] && [[ -f "$nbar_angle" ]] ; then
   echo "Running NBAR"
   cp "$nbar_input" "$griddedoutput"
   derive_l8nbar "$nbar_input" "$nbar_angle" "$nbar_cfactor"
